Refactor: Update home page layout for responsiveness
This commit refactors the home page layout to be more responsive. It changes the main `div` from a simple flex container to a flex container that switches from column to row layout on medium-sized screens and up. Also the side bar width uses `w-full md:w-64` to fill the screen on smaller screens.